changed parameter to one variable
Searching for hidingspot range is now always less far than the range when the automatic callback is done.

Also the range for hiding spot is now determined from the distance of the player and not the brother.

But determing if a spot is great depending on the distance is still relative to brother pos. So brother will still give greater points to hiding spot relative to his distance to the spot.